pyproject.toml
[project]
name = "led-controller"
version = "0.1.0"
description = "This is a simple LED controller library by leveraging the ELK-BLEDOM Bluetooth interface. It allows you to control the brightness and color of your LEDs using Python."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleak==1.1.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "schedule==1.2.2",
]
